feat: add restaurant creation & remove repository abstraction wrapper around sqlc

+++ internal/database/migrations/20250424132614_create_restaurant_users_table.sql
@@ -4,7 +4,7 @@ CREATE TABLE restaurant_users (
   id SERIAL PRIMARY KEY,
   restaurant_id UUID NOT NULL REFERENCES restaurants(id) ON DELETE CASCADE,
   user_id UUID NOT NULL REFERENCES users(id) ON DELETE CASCADE,
-  role_id INT NOT NULL REFERENCES roles(id)
+  role_id SMALLINT NOT NULL REFERENCES roles(id)
 );
 -- +goose StatementEnd
